app/backend/requirements.txt
fastapi==0.116.1
httpx==0.28.1
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.35.0
